small hzn changes and enhancements and merged Makefile
