+++ cli/hzn.go
@@ -3,12 +3,17 @@ package main
 
 import (
 	"github.com/open-horizon/anax/cli/cliutils"
-	"github.com/open-horizon/anax/cli/importkey"
 	"github.com/open-horizon/anax/cli/register"
-	"github.com/open-horizon/anax/cli/show"
 	"github.com/open-horizon/anax/cli/unregister"
 	"gopkg.in/alecthomas/kingpin.v2"
 	"os"
+	"github.com/open-horizon/anax/cli/key"
+	"github.com/open-horizon/anax/cli/node"
+	"github.com/open-horizon/anax/cli/agreement"
+	"github.com/open-horizon/anax/cli/metering"
+	"github.com/open-horizon/anax/cli/attribute"
+	"github.com/open-horizon/anax/cli/service"
+	"github.com/open-horizon/anax/cli/workload"
 )
 
 func main() {
@@ -19,52 +24,65 @@ func main() {
 	registerCmd := app.Command("register", "Register this edge node with Horizon.")
 	nodeIdTok := registerCmd.Flag("node-id-tok", "The Horizon exchange node ID and token. The node ID must be unique within the organization. If not specified, the node ID will be created by Horizon from the machine serial number or fully qualified hostname. If the token is not specified, Horizon will create a random token. If node resource in the exchange identified by the ID and token does not yet exist, you must also specify the -u flag so it can be created.").Short('n').PlaceHolder("ID:TOK").String()
 	userPw := registerCmd.Flag("user-pw", "User credentials to create the node resource in the Horizon exchange if it does not already exist.").Short('u').PlaceHolder("USER:PW").String()
-	inputFile := registerCmd.Flag("input-file", "A JSON file that sets or overrides variables needed by the node, workloads, and microservices that are part of this pattern. See /usr/horizon/samples/input.json. Specify -f- to read from stdin.").Short('f').String()
+	email := registerCmd.Flag("email", "Your email address. Only needs to be specified if: the node resource does not yet exist in the Horizon exchange, and the user specified in the -h flag does not exist, and you specified is the 'public' org. If all of these things are true we will create the user and include this value as the email attribute.").Short('e').String()
+	inputFile := registerCmd.Flag("input-file", "A JSON file that sets or overrides variables needed by the node, workloads, and microservices that are part of this pattern. See /usr/horizon/samples/input.json and /usr/horizon/samples/more-examples.json. Specify -f- to read from stdin.").Short('f').String()
 	org := registerCmd.Arg("organization", "The Horizon exchange organization ID.").Required().String()
 	pattern := registerCmd.Arg("pattern", "The Horizon exchange pattern that describes what workloads that should be deployed to this node.").Required().String()
 
-	importkeyCmd := app.Command("importkey", "Import a public key to verify signed microservices and workloads.")
-	keyFile := importkeyCmd.Arg("keyfile", "The public PEM file.").Required().String()
+	keyCmd := app.Command("key", "List and manage keys for signing and verifying services.")
+	keyListCmd := keyCmd.Command("list", "List the signing keys.")
 
-	showCmd := app.Command("show", "Display information about this Horizon edge node.")
-	showNodeCmd := showCmd.Command("node", "Show general information about this Horizon edge node.")
-	showAgreementsCmd := showCmd.Command("agreements", "Show the active or archived agreements this edge node has made with a Horizon agreement bot.")
-	archivedAgreements := showAgreementsCmd.Flag("archived", "Show archived agreements instead of the active agreements.").Short('r').Bool()
-	showMeteringCmd := showCmd.Command("metering", "Show metering (payment) information for the active or archived agreements.")
-	archivedMetering := showMeteringCmd.Flag("archived", "Show archived agreement metering information instead of metering for the active agreements.").Short('r').Bool()
-	showKeysCmd := showCmd.Command("keys", "Show the public signing keys that have been imported to this Horizon edge node.")
-	showAttributesCmd := showCmd.Command("attributes", "Show the global attributes that are currently registered on this Horizon edge node.")
-	showServicesCmd := showCmd.Command("services", "Show the microservices that are currently registered on this Horizon edge node.")
-	showWorkloadsCmd := showCmd.Command("workloads", "Show the workloads that are currently registered on this Horizon edge node.")
+	nodeCmd := app.Command("node", "List and manage general information about this Horizon edge node.")
+	nodeListCmd := nodeCmd.Command("list", "Display general information about this Horizon edge node.")
+
+	agreementCmd := app.Command("agreement", "List or manage the active or archived agreements this edge node has made with a Horizon agreement bot.")
+	agreementListCmd := agreementCmd.Command("list", "List the active or archived agreements this edge node has made with a Horizon agreement bot.")
+	listArchivedAgreements := agreementListCmd.Flag("archived", "List archived agreements instead of the active agreements.").Short('r').Bool()
+	agreementCancelCmd := agreementCmd.Command("cancel", "Cancel 1 or all of the active agreements this edge node has made with a Horizon agreement bot. Usually an agbot will immediately negotiated a new agreement. If you want to cancel all agreements and not have this edge accept new agreements, run 'hzn unregister'.")
+	cancelAllAgreements := agreementCancelCmd.Flag("all", "Cancel all of the current agreements.").Short('a').Bool()
+	cancelAgreementId := agreementCancelCmd.Arg("agreement-id", "The active agreement to cancel.").String()
+
+	meteringCmd := app.Command("metering", "List or manage the metering (payment) information for the active or archived agreements.")
+	meteringListCmd := meteringCmd.Command("list", "List the metering (payment) information for the active or archived agreements.")
+	listArchivedMetering := meteringListCmd.Flag("archived", "List archived agreement metering information instead of metering for the active agreements.").Short('r').Bool()
+
+	attributeCmd := app.Command("attribute", "List or manage the global attributes that are currently registered on this Horizon edge node.")
+	attributeListCmd := attributeCmd.Command("list", "List the global attributes that are currently registered on this Horizon edge node.")
+
+	serviceCmd := app.Command("service", "List or manage the microservices that are currently registered on this Horizon edge node.")
+	serviceListCmd := serviceCmd.Command("list", "List the microservices that are currently registered on this Horizon edge node.")
+
+	workloadCmd := app.Command("workload", "List or manage the workloads that are currently registered on this Horizon edge node.")
+	workloadListCmd := workloadCmd.Command("list", "List the workloads that are currently registered on this Horizon edge node.")
 
 	unregisterCmd := app.Command("unregister", "Unregister and reset this Horizon edge node so that it is ready to be registered again. Warning: this will stop all the Horizon workloads running on this edge node, and restart the Horizon agent.")
 	forceUnregister := unregisterCmd.Flag("force", "Skip the 'are you sure?' prompt.").Short('f').Bool()
 	removeNodeUnregister := unregisterCmd.Flag("remove", "Also remove this node resource from the Horizon exchange (because you no longer want to use this node with Horizon).").Short('r').Bool()
 
-	app.Version("0.0.2") //todo: get the real version of anax
+	app.Version("0.0.3") //todo: get the real version of anax
 
 	// Decide which command to run
 	switch kingpin.MustParse(app.Parse(os.Args[1:])) {
 	case registerCmd.FullCommand():
-		register.DoIt(*org, *pattern, *nodeIdTok, *userPw, *inputFile)
-	case importkeyCmd.FullCommand():
-		importkey.DoIt(*keyFile)
-	//case showCmd.FullCommand():   // <- I'd like to just show usage for hzn show, but don't know how to do that yet
-	//	showCmd.?
-	case showNodeCmd.FullCommand():
-		show.Node()
-	case showAgreementsCmd.FullCommand():
-		show.Agreements(*archivedAgreements)
-	case showMeteringCmd.FullCommand():
-		show.Metering(*archivedMetering)
-	case showKeysCmd.FullCommand():
-		show.Keys()
-	case showAttributesCmd.FullCommand():
-		show.Attributes()
-	case showServicesCmd.FullCommand():
-		show.Services()
-	case showWorkloadsCmd.FullCommand():
-		show.Workloads()
+		register.DoIt(*org, *pattern, *nodeIdTok, *userPw, *email, *inputFile)
+	case keyListCmd.FullCommand():
+		key.List()
+	//case keyCmd.FullCommand():   // <- I'd like to just default to list in this case, but don't know how to do that yet
+	//	keyCmd.List()
+	case nodeListCmd.FullCommand():
+		node.List()
+	case agreementListCmd.FullCommand():
+		agreement.List(*listArchivedAgreements)
+	case agreementCancelCmd.FullCommand():
+		agreement.Cancel(*cancelAgreementId, *cancelAllAgreements)
+	case meteringListCmd.FullCommand():
+		metering.List(*listArchivedMetering)
+	case attributeListCmd.FullCommand():
+		attribute.List()
+	case serviceListCmd.FullCommand():
+		service.List()
+	case workloadListCmd.FullCommand():
+		workload.List()
 	case unregisterCmd.FullCommand():
 		unregister.DoIt(*forceUnregister, *removeNodeUnregister)
 	}
