--- cli/importkey/importkey.go
@@ -1,7 +1,0 @@
-package importkey
-
-import "fmt"
-
-func DoIt(keyFile string) {
-	fmt.Println("Not implemented yet.") //todo: implement
-}
